requirements.txt
torch~=2.4.1
numpy~=1.24.3
scipy~=1.12.0
torchvision~=0.19.1
pillow~=10.2.0
setuptools~=68.2.0
pandas~=2.1.4
joblib~=1.3.2
scikit-learn~=1.4.1.post1
six~=1.16.0
matplotlib~=3.8.3
requests~=2.32.3
telegram~=0.0.1
pydantic~=2.8.2
aiogram~=3.12.0
fastapi~=0.114.0
python-multipart
uvicorn~=0.30.6
dotenv~=0.0.5
fairseq~=1.0.0a0+1254f60
bitarray~=2.9.2
omegaconf~=2.0.6
tensorboardX~=2.6
sacrebleu~=2.4.3
tqdm~=4.66.5
soundfile~=0.12.1
tabulate~=0.9.0
regex~=2024.7.24
wget~=3.2
librosa~=0.10.0
PyYAML~=6.0.2
iopath~=0.1.10
huggingface-hub~=0.24.6
timm~=0.6.11
opencv-python~=4.7.0.72
einops~=0.6.1
annoy~=1.17.3
python-dotenv~=1.0.1